chore: refactor test and implement branch test suggestion
